add some columns to alloc profiler table
Summary: The logging happens after the request finishes; thus we can add total_alloc, mem_peak to the samples, to allow us to zoom in onto those huge requests. The previous "total_alloc" meant the number of bytes already allocated at the time of the sample, and it is renamed to "curr_alloc" after this.

Reviewed By: alexeyt

Differential Revision: D20325402

fbshipit-source-id: a813d4f9d5c742d2afe3d5fa57ceb56e461dcdf2