Drop copy arg from lval methods
Summary:
Most callers pass cowCheck() as the copy parameter to the lval methods. For those few callers that don't do this, the callers pass "false", because they know for one reason or another cowCheck would return false. We drop this parameter and call cowCheck inside the method instead, to slightly simplify this interface.

To restate this invariant: in the three places that we passed "false" for the copy parameter (two in APCLocalArray and one in RecordArray), calling "cowCheck()" within the helper would also return false. In the APCLocalArray case, that's because it's an invariant of APCLocalArray that the wrapped array always has refcount 1. In the RecordArray case, it's because we've just allocated the new MixedArray. As a result, this change has no effect on any COW behavior.

Furthermore, the attempt to optimize this check away by passing in the value is actually counter-productive - in all hot call-sites of lval, we need to call cowCheck() anyway, so setting up the argument register is an unnecessary instruction (since the method can do the test in memory). It doesn't really matter, because lval is cold.

Reviewed By: alexeyt

Differential Revision: D20896642

fbshipit-source-id: e694502e27528bf9536ec781e944fdda473044b5