Simplify ArrayKinds now that lval throws on missing keys
Summary:
There was a time in PHP when ArrayData::lval could create new keys. That time is now over, thanks to mxw eliminating "false-y promote".

Simplify logic in the codebase that dealt with the possibility of promotion, and in particular, eliminate code that escalated arrays prior to throwing. We could have split up this change by ArrayKind, but the three cases here are similar enough that they can go together.
- In EmptyArray, lval can now simply throw.
- In PackedArray's lval methods, we don't need to allocate a MixedArray and let it throw - we can throw ourselves.
- In MixedArray, we no longer need the addLvalImpl<enforce = true> case. As a result, we don't need to ensure that we have space for the new array.

This diff has a minor memory effect given the new implementation of MixedArray's lval methods: when the MixedArray has exactly 3 elements, we no longer double it to size 6 on a call to lval. Perf tests show that this change doesn't matter.

Reviewed By: alexeyt

Differential Revision: D20906202

fbshipit-source-id: c76020165f61d385a5eafd4f80981d9adaee9253