Add NamespaceDeclarationHeader node
Summary:
The direct decl parser is a full-fidelity parser SmartConstructors implementation which directly produces decls instead of going through the historical path of CST -> AST -> NAST -> decl. This means that it must emulate the behavior of the lowerer and Naming for type and function signatures. One of the behaviors of Naming that it must implement is fully qualifying the name of every type.

The structure of the CST means that this job is difficult in files containing namespace declarations. We must know what namespace we are in so that we can fully-qualify names in types, but the parser does not signal to us that we are in a namespace until we leave it. Consider:

```
namespace Foo {
  type A = int; // fully-qualified: type \Foo\A = int;
  type B = A;   // fully-qualified: type \Foo\B = \Foo\A;
  ...
}
```

The direct decl parser's smart constructors will be invoked by the full-fidelity declaration parser in this order (eliding a lot of detail for clarity):

```
make_token("namespace");
make_token("Foo");
make_token("{");

make_token("type"); make_token("A"); make_token("="); make_token("int"); make_token(";");
make_alias_declaration(Token("type"), Token("A"), Token("="), Token("int"), Token(";"));

make_token("type"); make_token("B"); make_token("="); make_token("A"); make_token(";");
make_alias_declaration(Token("type"), Token("B"), Token("="), Token("A"), Token(";"));

...

make_token("}");
make_namespace_body(
  Token("{"),
  List(vec![
    Typedef { .. }, // A = int
    Typedef { .. }, // B = A
    ...
  ]),
  Token("}"),
);
make_namespace_declaration(
  Token("namespace"),
  Token("Foo"), // Only now do we know what namespace we were in!
  NamespaceBody(...),
);
```

Our implementation of make_namespace_declaration is not invoked until well after we have constructed the Typedefs, so we were not able to fully-qualify them. We could rewrite all the decls upon invocation of make_namespace_declaration, but we want to avoid needing to traverse and mutate all the types in the namespace body.

To work around this problem, we added behavior to `make_token`, which watches for the `namespace` keyword, then adds the name following it to the current namespace stack. This works okay, but it conflicts with another bit of syntax, `use namespace`, which imports a namespace rather than declaring one. As a result, the direct decl parser incorrectly interpreted `use namespace` as declaring a namespace.

We might be able to continue hacking around the problem in the direct decl parser by adding more parsing logic to `make_token`, but I think it would be much simpler to change the CST. Suppose we added a namespace_declaration_header node, and constructed it before parsing the namespace body:

```
make_token("namespace");
make_token("Foo");
make_namespace_declaration_header(Token("namespace"), Token("Foo")); // Signal!
make_token("{");
// ...etc
```

Then it would be a child of the namespace_declaration node:

```
make_namespace_declaration(
  NamespaceDeclarationHeader(Token("namespace"), Token("Foo")),
  NamespaceBody(...),
);
```

And we would get a signal that we were entering a namespace before actually parsing its contents. This solution wouldn't interfere with namespace-use declarations.

This diff adds the NamespaceDeclarationHeader node to the CST, parses it in the FFP, lowers it, and handles it in CST-error-detection, the facts parser, and hackfmt. The next diff will revise the direct decl parser to make use of make_namespace_declaration_header instead of its make_token-based workaround.

Differential Revision: D20746539

fbshipit-source-id: 804ebb28cbff44452259923e2f624a978d92dfec