Unbox Typing_deps.variant
Summary:
jakebailey pointed out that, as written, we do an extra allocation for the tuple cases of this variant. We can avoid it by not wrapping the tuple in parentheses (!).

Since they're not boxed anymore, we have to explicitly pattern-match against them in `variant_equals`, instead of binding the entire tuple to a value.

Differential Revision: D20882290

fbshipit-source-id: 00775e4a303570b9491090f30be9a591dcc48697