Report duplicate name errors in `hh_lite`
Summary:
These are reported as part of updating the global reverse naming table, rather than as part of the `Naming.program` call.

Arguably, the `Naming.program` call should be responsible for that, and updating the reverse naming table should never produce errors (and instead be a multimap). But that would require some refactoring.

Reviewed By: ljw1004

Differential Revision: D19277783

fbshipit-source-id: 3157dd400c1d3d8ca8f48e85bead9ff204f20619