Improve formatting of format string error messages
Summary:
Current string error messages are not embraced by double quotes and this will cause confusions when there is a `%` in the string.
For example, when the user write:
```
<?hh

function foo(): void {
  $s = Str\format("hello %");
}
```
`hh` will produce this error message:
```
Typing[4038] Invalid format string % in hello %
```
but what we want is:
```
Typing[4038] Invalid format string % in "hello %"
```
This diff solves this problem by changing the `errors.ml` file and another test is changed accordingly in order to land this diff.

Reviewed By: Wilfred

Differential Revision: D20927362

fbshipit-source-id: 00e7944cd86455b52c3b2518df0afc4c0d298d41