Fix typed native builtins which take param as const Variant&
Summary:
Its possible to have a builtin take a parameter as const Variant&, but
its typed in PHP as taking something other than mixed. These params
will have a builtinType, even though they expect a Variant ref. For
ref types, right now we pass just a pointer to the value to the
builtin (like we would for Array or String etc). This works because
passing a pointer to the value happens to be exactly the same as
passing a pointer to the whole TypedValue.

This breaks when the value pointer isn't coming from a TypedValue. Add
two flags to ParamInfo, indicating that the native builtin takes the
param as a const Variant& or a TypedValue. This is determined and set
in func-emitter.cpp when creating the Func. If either flag is set,
when emitting code to call the builtin in, pass it properly as a
TypedValue or const Variant& regardless of what its builtinType is.

Reviewed By: jano

Differential Revision: D20912123

fbshipit-source-id: e91a1ebcd127378220357a6d9dc5a0604e643688