Support pocket universes syntax
Summary:
Add support for the PU syntax. The formatting rules implemented here are pretty straightforward--we may want to do further work here. For instance, we might want to allow collapsing a PocketAtomMappingDeclaration onto a single line when it fits, like this:

```
  enum E {
    case int v;
    :X(v = 1);
    :Y(v = 2);
  }
```

Rather than always expanding it, as this diff does:

```
  enum E {
    case int v;
    :X(
      v = 1
    );
    :Y (
      v = 2
    );
  }
```

But maybe that's an improvement we can leave to a future diff.

Reviewed By: vsiles

Differential Revision: D20928501

fbshipit-source-id: ead54f91af727ea6f5b21382be261f316ab48f84