Replace use of bools in param by bitwise flags
Summary:
The `FunParam` type is currently defined as follows:
```
type 'ty fun_param = {
  fp_pos: Pos.t;
  fp_name: string option;
  fp_type: 'ty possibly_enforced_ty;
  fp_kind: param_mode;
  fp_accept_disposable: bool;
  fp_mutability: param_mutability option;
  fp_rx_annotation: param_rx_annotation option;
}
```
This is 8 words, or 64 bytes, for every parameter in a function type. Three of these fields are booleans or small enums, and can instead be encoded using bit flags, saving two words, reducing the size of the object to 48 bytes.

Differential Revision: D20913265

fbshipit-source-id: 9f9bf97fa2aeea58bd0d97e9ea2cf5248399c306