Add definition site position to cyclic typedef error
Summary:
The reporting of cyclic type defs uses a position derived from reason information in a type as the primary position.

Instead, report the definition itself as the primary position, and report the particular cyclic use as a secondary position. This is a slightly more informative message; it's also another step towards the goal of eliminating reasons in types as sources for primary error positions.

Reviewed By: Wilfred

Differential Revision: D20892945

fbshipit-source-id: e0cf5a141ad0ce7c237f6311bf41ff8b0f303e32