Drop the APCLocalArray optimization
Summary:
To store array-likes in APC, we attempt to convert them to uncounted arrays, which may fail, e.g. if the array-like (transitively) includes an object. In these cases, we instead create an APCArray to represent them and store that.

A "plain array" is a PHP array that's not a varray or darray. Plain arrays are quite rare in production code today. To materialize a plain array from an APCArray, we currently create an APCLocalArray wrapper type around it that supports the ArrayKind interface. By contrast, for varrays, darrays, and Hack arrays, we do a full copy. With this change, we'll do the same copy for plain arrays, so we no longer need APCLocalArray.

This code is cold because triggering this code path requires storing a) a plain array with b) objects inside it in APC, which just doesn't happen anywhere. In fact, there's no test coverage at all, so I added one.

Thanks alexeyt and ricklavoie for the suggestion.

Reviewed By: mxw

Differential Revision: D20942877

fbshipit-source-id: 078bf389091ee0bf6728cdbcdd2e3f9f69ee6a49