track hh_server_status_diagnostic
Summary:
My plan is to report hh_server_status via a diagnostic.

Therefore clientLsp has to track which status diagnostic it has already sent out. It needs this so that it can dismiss or update the diagnostic once things change.

That's all this diff does. It merely adds tracking for it, and ensures that (if created) it gets disposed of at the right times. It doesn't however ever create the status-diagnostic; that's left for a future diff.

I made a simplifying decision that status-diagnostics will only ever be reported during In_init and Lost_server states. That's because in Main_loop state then we know there's no status to report! The decision also has the advantage that In_init and Lost_server states never report any diagnostics from hh_server, and therefore there's never any need to merge status-diagnostic with normal diagnostics.

Reviewed By: arxanas

Differential Revision: D20877674

fbshipit-source-id: 4ed5a192d96d57be3f8dd9237891e4c9039f5911