support status diagnostics
Summary:
I wish to support sending status-about-diagnostics over publishDiagnostics.

That's fine, and the LSP spec will happily let me do it as-is.

But for testing purposes, it'd be useful for tests to distinguish which publishDiagnostics were carrying true diagnostics, and which were merely carrying status.

This diff adds a new FB-specific flag "isStatusFB" to the PublishDiagnostics.params to indicate that fact. It's a boolean, and can be omitted (which means the same as 'false')

This diff alters the LSP test framework to support ignoring these status diagnostics.

Reviewed By: arxanas

Differential Revision: D20896162

fbshipit-source-id: 51f217c4424b2bcab4361d4008b69362fecc25f7