publish hh_server_status as a diagnostic
Summary:
This diff reports hh_server status as a diagnostic.

(Note that hh_status currently also reports its status with window/showStatus. That's addressed with the next diff in the stack).

It reports status in the file that's currently active in the IDE, to be always prominent.

I made the decision that if no hack files are open in the IDE, then it won't report status. That's for folks like Flow developers who inadvertently open a .php file, but close it again and don't care about hh_server at all, and don't want the hh_server message to forever pollute their diagnostics list.

Note: as discussed in an earlier diff, LSP and VSCode are only able to show diagnostics attached to a particular file (so we can't display hh_server status diagnostic just as a general unattached diagnostic).

Moreover, LSP requires every diagnostic to be attached to a range within that file. VSCode has a few conventions to e.g. highlight the first word on a line, or an entire line, but they're not supported by other editors:
https://github.com/microsoft/language-server-protocol/issues/257

That's why I decided to report the error only on the first character of the file. (it's what arc lint does too).

Reviewed By: arxanas

Differential Revision: D20896184

fbshipit-source-id: 3817ba187d52650e73ee2349b0e4771bb7d95885