normalize status display
Summary:
This diff provides a systematic overhaul of the messages displayed in the status bar.

The principle for the status-bar text is this:
* "Hack" in the status bar means that IDE services are working
* "Hack: busy" or "Hack: initializing" or "Hack: failure" meanss that IDE services aren't working for whatever reason. This is true whether IDE is powered by serverless-IDE or powered by hh_server

The principle for the status-bar icon is this:
* [ok] means that HackIDE and hh_server are both ready
* [spinner] means that HackIDE or hh_server are doing work
* [error] means that IDE services are denied due to an error.

Note for instance that if HackIDE is up but hh_server is down, then that gets shown as "[ok] Hack". That's because the fact of hh_server not working doesn't hinder IDE services, and nothing is busy processing, and the fact that we're not seeing diagnostics is happily represented by a status-diagnostic and doesn't also need to be shown in the status bar.

Reviewed By: arxanas

Differential Revision: D20886138

fbshipit-source-id: f61dfc8a32433848ad772458ff2cfd8abb52fd47