track most recent file
Summary:
My goal is to report hh_server status in a diagnostic, not in the status bar.

This will squarely address complaints of users who say "I never realized that hh_server wasn't reporting errors". It will also make the status bar cleaner.

VSCode and LSP don't support "general" diagnostics. They only support diagnostics attached to a real file:
https://github.com/microsoft/language-server-protocol/issues/256

I therefore decided that I will attach the status-diagnostic about "hh_server not running" to the most recently used file, if it's open.

This diff is in service of that goal. It simply tracks which was the file most recently used. There's no guarantee that the file is still open or anything. It just tracks the most recent file.

Reviewed By: arxanas

Differential Revision: D20876768

fbshipit-source-id: 7328c163212fd9d2309480d25713d768af6cf11a