show status promptly
Summary:
Previously, we updated status only upon idle tick. That happened if no other server/client events had happened for a second.

The effect was that, when you open a file, status didn't reflect that for a second or two (as the file got decl'd)

This diff changes it so status is updated as promptly as posssible. I figure that prompt status updates will make things feels snappier.

The refresh_status function declines to send any LSP status updates if status hasn't changed, so this change won't cause a flurry of updates.

One concern is during hh_server initialization, i.e. the In_init state. In that state we display the number of seconds that have elapsed. That won't update more than once a second, by definition, so we're fine.

The other concern is during sIDE processing_files, e.g. if the user did a rebase and there are 10k files to process. The sIDE will send a "process_file" update upon every single one of them. What I've done is made it so our status message only shows the total number of files that are being processed. This total number won't change frequently.

Reviewed By: arxanas

Differential Revision: D20901003

fbshipit-source-id: c3e35abfdbf69437bdf5344dbecb152e8a94d1cb