Allow cow/escalation in ArrayData::renumber
Summary:
All current ArrayKind implementations have the following property: the "renumber" operation can be performed on them without changing their layout.

We don't necessarily want to provide this property for bespoke array layouts, though. For example, we may have a record-like layout with integer keys; renumbering such a layout may require escalation to a MixedArray.

Update Renumber to allow for cow/escalation. There's actually a bug in the current implementation: we don't cow in MixedArray when we should. This bug is asymptomatic but only due to highly non-local invariants: the small handful of places that use Array::merge (which calls renumber) happen to provide it EITHER an empty array or an array with refcount 1. I hate having to do such a complex proof of correctness, so I just had MixedArray::Renumber do the cow check as well.

The other ArrayKind implementations never need to cow or escalate, because Renumber is trivial on them. (Preserving PHP behavior, it has no effect on the globals array.)

Reviewed By: ricklavoie

Differential Revision: D20952787

fbshipit-source-id: d69ca22e5a6531022e7ea2b848cd6a88136dc9bf