better logging of failures
Summary:
I noticed that our telemetry logs show a bunch of failures where data is simply the string "NoData". I also saw that our LSP transcripts were failing to show enough information about some errors. This diff is an end-to-end push to make record better errors.

The root cause is an `LspException` being thrown with `data=None`, and there's no place in the codebase that adds any data. (I think this is a flaw which crept in when I added the ability for the code to throw LspExceptions so it could decide which error code and message to use. I failed to populate the stack).

1. This diff adds a callstack to exceptions if one wasn't already present. (There might be one already. For instance, when we receive an exception message from hh_server, then that message includes the remote stack, which is the oen we use).
2. In places In_init and Lost_server we have catch-all message handlers that emit `data=None`. I think it'd be useful for those handlers to at least indicate which message they're responding to. In the case of notifications, the information would be otherwise completely lost.
3. In cases where we want to report an error over LSP for an event that wasn't a request, well, jsonrpc offers no way to do that. So we've been rolling our own with telemetry/event. Prior to this diff we'd been sticking everything into just a string. This diff restores structure to our telemetry/event.
4. I made it also include message filename in `Lsp_fmt.denorm_message_to_string` since filename is handy in seeing at-a-glance what a message is for.

Reviewed By: 2BitSalute

Differential Revision: D20951425

fbshipit-source-id: a69d28bf57b1c753ca84b3ee6349d6d8e7c07ead