ServerConfig.load ~silent
Summary:
Normally when a process loads server-config, it dumps its contents plus overrides to stderr.

This is fine for hh_server startup.

But it ends up getting dumped twice for LSP, once for the clientLsp process and once for the clientIdeDaemon process. And actually it shouldn't be dumped at all since it's wrong for us to write to stderr when the LSP process is connected to VSCode (there's no one to evenn read it). And also I've never found it useful to see this dump of hhconfig options.

So I'm threading through the `~silent` option, and disabling it from clientLsp and clientIdeDaemon.

Reviewed By: arxanas

Differential Revision: D20959800

fbshipit-source-id: eaab3ce8f17e196c1e1aa276bf5f85fef0d5f549