Fix formatting of empty pocket atom mappings list
Summary:
`braced_block_nest` was used for this construct, but it assumes that the delimiters will always be present, making it behave incorrectly for an empty pocket atom mappings list. It unconditionally adds a newline, producing:

```
class PU1 {
  enum X {
    :X
    ;
  }
}
```

Instead, do not add a newline when there are no mappings and the opening paren is missing.

Reviewed By: vsiles

Differential Revision: D20975419

fbshipit-source-id: a84d2d7d3a3a99c272532f0f2d928e89037eb296