Fix key refcounting bug
Summary:
I introduced this bug in D20820640. When I audited users of MixedArrayElm::getKey, nvGetKey, nvFirst, and nvFirstLocal, I missed this site.

Generally speaking, if a call-site of one of these methods does the following, it isn't assuming an inc-ref of the key (and the old inc-ref behavior will actually cause it to leak):
```
for (ArrayIter it(arr); ...) {
  auto const key = it.nvFirst();
  other->set(key, value);
}
```

These cases were the majority of call-sites. In this particular example, I missed that following the set, there's a dec-ref of the key which is no longer needed.

While re-reading all call-sites I found a few related incorrect comments, which I've also fixed.

Reviewed By: jjgriego

Differential Revision: D20984289

fbshipit-source-id: 91016bbfb03da29fe548c87968c1ea838eaf63c2