Setting for parallel type checking threshold
Summary:
The default value is identical to the current behavior: when there are fewer than 10 files to recheck, we recheck them in the master process without using the MultiWorker parallelism.

Adding this setting should enable us to turn off sequential execution for certain scenarios, and perhaps for all scenarios at some point in the future.

We've previously discussed whether it ever makes sense to process files sequentially, and there didn't seem to be any evidence to support that.

Reviewed By: yinghuitan

Differential Revision: D20958765

fbshipit-source-id: 3bdf94952c7984fdc0d18201b739d98cf2c6cc0c