Implement ToOcamlRep for str, [u8], and [T] where T: ToOcamlRep
Summary: Now that the previous diff makes it possible to implement ToOcamlRep for unsized types, we can add implementations for the unsized versions of a few types that we already convert. Namely, `str` (which is converted to `string`), `[u8]` (converted to `Caml.Bytes.t`, which is equal to `string`), and [T] where `T : ToOcamlRep` (converted to `list`).

Reviewed By: dtolnay

Differential Revision: D20206584

fbshipit-source-id: 24a167183283e0ae415875a49cc7e15a7e336218