Split OcamlRep trait into ToOcamlRep and FromOcamlRep
Summary:
As noted in T62210650, it's currently possible to write `arena.add(&"...".to_owned())` and `arena.add(&Cow::Borrowed("..."))`, but not `arena.add("...")`. The current API requires consumers to use `str_to_ocamlrep` instead. This is because OcamlRep requires the Sized bound--`from_ocamlrep` returns `Self`, and we cannot return an owned value of an unsized type. So because `str: !Sized`, we cannot implement `OcamlRep` for it as long as OcamlRep requires `Sized`.

Because only `from_ocamlrep` (and `from_ocaml`) involve handling an unboxed value (of the type implementing OcamlRep), we can separate OcamlRep into two traits (ToOcamlRep and FromOcamlRep), and only require Sized on FromOcamlRep.

In addition to allowing unsized types to implement ToOcamlRep, we can statically prevent a few panics. We have several types which we have only needed to convert to OCaml, and for which we just panic in `from_ocamlrep`. With the separate trait, we can properly express statically that there is no OCaml->Rust conversion for these types.

Reviewed By: losvald

Differential Revision: D20955458

fbshipit-source-id: eda1e64d05f7244878cdf36dfb834c0bdbce1830