Improve line numbers when suspended in concurrent.
Summary:
When we lower concurrent blocks to Awaitall nodes, lower the position as that of the keyword, and emit the Awaitall nodes with that position on the shared code we suspend in. Add test that verifies that multiple frames resuming while their parent is the concurrent block will have the line number of the concurrent keyword (not that of the lexicographically last lifted await) in their backtraces.

We also lower await as expression to Awaitall nodes, so make the test verify that backtraces through those don't change: they still point to the end of the statement.

While we're in the lowerer, fix "scrope" typo.

Reviewed By: oulgen

Differential Revision: D21037266

fbshipit-source-id: 67b6e56ff8f7481d7769616431069014229a6df1