Add `Value::as_byte_string`
Summary: The `Value::as_str` API returns a `Option<Cow<str>>`. When we're trying to hash the string, we don't need to do an extra iteration over the string to confirm that it's valid UTF-8, since we don't actually care about the encoding. This commit adds an API to get to the bytes underlying the string for this purpose.

Differential Revision: D21010468

fbshipit-source-id: afe8b8e5e11e6870b62814aaeac8cd7b608ef13c