switch to 'ClientIdeService.t option'
Summary:
Previously we said that clientIdeService.t always existed, and it pumps messages if env.use_serverless_ide, and we're supposed to check if 'env.use_serverless_ide' to know whether to send messages to it. (the penalty of sending messages when it's not pumping is an indefinitely growing queue).

This discipline held for notify_disk_file_changed because we only registered for DidChange messages if env.use_serverless_ide. It didn't help for notify_ide_file_open/close. I think the discipline wasn't easy to spot.

This diff therefore changes it to a straightforward rule "ide_service is a option". If it's Some, then you can use it.

Reviewed By: arxanas

Differential Revision: D21018109

fbshipit-source-id: 8366d74cf07ea8a5f16f2107d4efc40cd35da950