we can assert that initializeParams are present in server-status
Summary:
I'm working through the uses of initialize_params_exc. This one is weird. If initializeParams hasn't already been recieved by this stage the something's wrong.

Context: I think that the initialize-params promise is racey.
* Imagine we receive "initialize" from the client, which calls `Lwt.wakeup_later initialize_params_resolver`, which means that the pending call to `let%lwt root = get_root_wait ()` inside run_ide_service is able to proceed
* Imagine we immediately also receive "DidOpen" from the client, which means that `let%lwt event = get_next_event` is able to proceed

Is there any guarantee that run_ide_service will proceed first? Or is it possible that the main message handler will proceed first, and encounter DidOpen, and try to send it to ClientIdeService, even before we've sent Initialize_from_saved_state to ClientIdeService?

The docs for `Lwt.wakeup_later` say "This triggers callbacks attached to the promise". It's not clearly specified whether it triggers them synchronously right here and now, or whether it merely triggers them in the sense that they're able to run and will definitely run soon. I think that in async libraries the former is rare because of the danger of blowing the stack in recursive async functions. Maybe ocaml would be okay thanks to tail recursion. It bears further investigation.

The source code documentation of Lwt says this https://github.com/ocsigen/lwt/blob/2412006f228622437f5625d6e71dcdd0ea5cbcb2/src/core/lwt.ml#L1108
> Current Lwt actually has a messy mix of callback-calling behaviors. For example, [Lwt.bind] is expected to always call its callback immediately, while [Lwt.wakeup_later] is expected to defer all callbacks of the promise resolved, {e unless} Lwt is not already inside the resolution loop.

Well, that's not a strong enough guarantee for us to build upon.

Therefore my project is to get rid of initialize_params_promise since awaiting on it is solely used in the current unreliable place.

Instead I'll send the Initialize_from_saved_state message to clientIdeService immediately in the "initialize" handler. I'll switch the promise so it's no longer waiting for initializeParams; instead it will wait for ide_service's initialize to be finished.

This will guarantee there are no races. We're relying on the fact that LSP guarantees that no messages are sent prior to initialize, and we specifically reject such messages with an error message should a client send them.

Reviewed By: arxanas

Differential Revision: D21031282

fbshipit-source-id: d39a93bb9d08bb8439b2501125849a27b7c32a90