use ref not promise for initialize_params global
Summary:
Previously we stored initialize_params behind a promise, so that people could be woken up once it was available.

There was only one place which used that feature, `run_ide_service`, and I had my doubts about it and removed it.

I think having a promise here is dangerous, because it makes people think they can usefully be triggered once the promise is fulfilled, but I think correctness of such wakeups is subtle and needs more care. By changing this global to be a ref not a promise, I think people won't feel so welcome to trigger on it.

Reviewed By: arxanas

Differential Revision: D21032564

fbshipit-source-id: e958935927b9feab0ca126b2ed6f9833b98114d3