do_rpc
Summary:
This diff just renames do_rpc to rpc_internal, and changes the caller of "do_rpc" into "rpc".

In this way, "rpc" is the only entry-point for folks doing rpc.

(This is a trivial diff. I'm never sure how well moving-blocks-of-code-around diffs like, so I try to do them on their own).

Reviewed By: arxanas

Differential Revision: D21024807

fbshipit-source-id: 0798ea8f3711278aa72280aee5ae1488f861e52d