Fix post-conditions for regions ending in FCall*s with asyncEagerReturn
Summary:
The callee return type is different in each of the two FCall*'s successors in this case.  The logic to collect the post-conditions looks for the types of the VM-state slots in the main exit of the region, which is the non-eager return. This prevents the arc from the FCall* to the eager-return block from being included in the optimized region in the cases where it guards on the function return type, because the pre/post-conditions mismatch.

The fix here is to avoid asserting the return type on the two exits following the FCall* in profiling translations. This avoids the incorrect post-conditions to be recorded for profiling translations, which then allows both paths to be included in a single optimized region.

Reviewed By: binliu19, paulbiss, jano, mofarrell

Differential Revision: D21066112

fbshipit-source-id: 11eeefc3404fe6f6c592b17453e7891bec846467