Side-exit on dvarray mismatch
Summary:
By doing a side exit on dvarray type hint failure, we can use the check to de-virtualize other array operations later in the translation. The guard already essentially takes care of this piece for param type checks, but not for return value checks (which are often not in a guardable position).

Specifically, this change helps when we inline a function which returns (say) a varray, and then the caller iterates over that varray - now we'll have layout information at that point.

Reviewed By: billf

Differential Revision: D21060258

fbshipit-source-id: b57e32af1e0bd89181e0e2d6237de05990e772ad