Updating `Members` to use `keyset` instead of `vec` + recursive call
Summary:
The current implementation only dumps the instances names of the current class,
and do not look for the one in the parents. By recursively calling the parent
class, we fix this.
Also using `keyset` to avoid dups.

Note: the parent detection is wrong (based on `extends`) and still under
discussion. Once we settle on a solution, I'll update the rust code directly
in the right version

Reviewed By: francesco-zappa-nardelli

Differential Revision: D20915436

fbshipit-source-id: fb11b00ca77bc3f6f7d2cdbac74acc61dffd1e66